requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
requests==2.31.0
beautifulsoup4==4.12.3
pydantic==2.7.1
google-generativeai==0.7.2
urllib3==2.2.1fast